fix thigs up

+++ database/migrations/2023_06_12_033317_create_invoice_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->integer('invoice_id')->unsigned();
             $table->text('nokw');
             $table->date('date');
-            $table->decimal('price', 8, 2);
+            $table->bigInteger('price');
             $table->timestamps();
 
             $table->foreign('invoice_id')->references('id')->on('invoices')->onDelete('cascade');
